Moved the 'profile.py' module down to the 'managers' folder. No change in functionality.
